requirements.txt
streamlit==1.32.0
pandas==2.2.1
openpyxl==3.1.2
openai==1.13.3
python-dotenv==1.0.1
numpy==1.26.4
crewai==0.11.0
tiktoken>=0.5.2,<0.6.0
langchain-core>=0.3.29
langchain-google-genai>=0.0.11
together>=0.2.8
